requirements.txt
python-dotenv==1.0.0
aiogram==3.2.0
sqlalchemy==2.0.23
python-telegram-bot==20.6
openai==1.3.0
langchain==0.0.335
fastapi==0.105.0
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
sqlmodel==0.0.11
